Fix: integer overflows calculating the number of pixels in imbuf
Use size_t for imbuf size calculation to prevent integer overflow.

Back ported from: 6d6c695619b5199d2306eede273b9c50ebd10a28

Pull Request: https://projects.blender.org/blender/blender/pulls/137443

+++ source/blender/imbuf/intern/rectop.cc
@@ -1239,11 +1239,11 @@ void IMB_rectfill_area(
 
 void IMB_rectfill_alpha(ImBuf *ibuf, const float value)
 {
-  int i;
+  size_t i;
 
   if (ibuf->float_buffer.data && (ibuf->channels == 4)) {
     float *fbuf = ibuf->float_buffer.data + 3;
-    for (i = ibuf->x * ibuf->y; i > 0; i--, fbuf += 4) {
+    for (i = size_t(ibuf->x) * size_t(ibuf->y); i > 0; i--, fbuf += 4) {
       *fbuf = value;
     }
   }
@@ -1251,7 +1251,7 @@ void IMB_rectfill_alpha(ImBuf *ibuf, const float value)
   if (ibuf->byte_buffer.data) {
     const uchar cvalue = value * 255;
     uchar *cbuf = ibuf->byte_buffer.data + 3;
-    for (i = ibuf->x * ibuf->y; i > 0; i--, cbuf += 4) {
+    for (i = size_t(ibuf->x) * size_t(ibuf->y); i > 0; i--, cbuf += 4) {
       *cbuf = cvalue;
     }
   }
